refactor: move useEffect logic in PermissionsList to its own hook
